reorg adhoc ui
